Move generally-useful functions to separate module
There is now a Language.C.Syntax.Utils module that contains useful
functions for dealing with ASTs. Right now, these include a
manually-written traversal to extract sub-statements and one to apply a
function to all sub-statements. These might be better written with some
sort of generic programming framework in the future.