Allow void functions to return void expressions
There really is code (in Linux) that looks like this:

void g();
void f() { return g(); }

The standard doesn't allow it (6.8.6.4.1) but GCC does, so it seems
useful to allow it ourselves, too.

Eventually, it would make sense to have some configuration settings that
describe exactly which variant of the language to accept. In strict C99
mode, we'd reject this.