Parse (and ignore) Clang __attribute__((availability(id=major.minor.rev)))
If the version is just major.minor, it will lex as a float literal.
If it's major.minor.rev it will lex as a new clang version token.

In either case the parser will throw those attribute parameters out. (As it
currently does with other 'unary_expr = unaray_expr' attribute parameters).